# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soundmark API",
    "description": "Location-based social music recommendation platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": [
          "Users & Follow"
        ],
        "summary": "Get User Profile",
        "description": "Get user profile with follow statistics\n\nArgs:\n    user_id: ID of the user to retrieve\n    current_user: Currently authenticated user (optional)\n\nReturns:\n    User profile with follower/following counts and relationship status",
        "operationId": "get_user_profile_api_v1_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}/stats": {
      "get": {
        "tags": [
          "Users & Follow"
        ],
        "summary": "Get User Follow Stats",
        "description": "Get user's follow statistics\n\nArgs:\n    user_id: ID of the user\n\nReturns:\n    Follower and following counts",
        "operationId": "get_user_follow_stats_api_v1_users__user_id__stats_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}/follow": {
      "post": {
        "tags": [
          "Users & Follow"
        ],
        "summary": "Follow User",
        "description": "Follow a user\n\nArgs:\n    user_id: ID of the user to follow\n    current_user: Currently authenticated user\n\nReturns:\n    Success status and updated follower count",
        "operationId": "follow_user_api_v1_users__user_id__follow_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users & Follow"
        ],
        "summary": "Unfollow User",
        "description": "Unfollow a user\n\nArgs:\n    user_id: ID of the user to unfollow\n    current_user: Currently authenticated user\n\nReturns:\n    Success status and updated follower count",
        "operationId": "unfollow_user_api_v1_users__user_id__follow_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}/followers": {
      "get": {
        "tags": [
          "Users & Follow"
        ],
        "summary": "Get User Followers",
        "description": "Get list of users following this user\n\nArgs:\n    user_id: ID of the user\n    limit: Maximum number of results (1-100)\n    offset: Number of results to skip\n\nReturns:\n    List of followers and total count",
        "operationId": "get_user_followers_api_v1_users__user_id__followers_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}/following": {
      "get": {
        "tags": [
          "Users & Follow"
        ],
        "summary": "Get User Following",
        "description": "Get list of users this user is following\n\nArgs:\n    user_id: ID of the user\n    limit: Maximum number of results (1-100)\n    offset: Number of results to skip\n\nReturns:\n    List of users being followed and total count",
        "operationId": "get_user_following_api_v1_users__user_id__following_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FollowResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "follower_count": {
            "type": "integer",
            "title": "Follower Count"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "follower_count"
        ],
        "title": "FollowResponse",
        "description": "Response for follow/unfollow actions"
      },
      "FollowStats": {
        "properties": {
          "follower_count": {
            "type": "integer",
            "title": "Follower Count"
          },
          "following_count": {
            "type": "integer",
            "title": "Following Count"
          }
        },
        "type": "object",
        "required": [
          "follower_count",
          "following_count"
        ],
        "title": "FollowStats",
        "description": "User follow statistics"
      },
      "FollowersResponse": {
        "properties": {
          "followers": {
            "items": {
              "$ref": "#/components/schemas/UserPublic"
            },
            "type": "array",
            "title": "Followers"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "followers",
          "total"
        ],
        "title": "FollowersResponse",
        "description": "Response for followers list"
      },
      "FollowingResponse": {
        "properties": {
          "following": {
            "items": {
              "$ref": "#/components/schemas/UserPublic"
            },
            "type": "array",
            "title": "Following"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "following",
          "total"
        ],
        "title": "FollowingResponse",
        "description": "Response for following list"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "spotify_id": {
            "type": "string",
            "title": "Spotify Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "spotify_id"
        ],
        "title": "UserPublic",
        "description": "Public user information (for followers/following lists)"
      },
...
      "UserWithStats": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "spotify_id": {
            "type": "string",
            "title": "Spotify Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "follower_count": {
            "type": "integer",
            "title": "Follower Count"
          },
          "following_count": {
            "type": "integer",
            "title": "Following Count"
          },
          "is_following": {
            "type": "boolean",
            "title": "Is Following",
            "default": false
          },
          "is_followed_by": {
            "type": "boolean",
            "title": "Is Followed By",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "spotify_id",
          "created_at",
          "follower_count",
          "following_count"
        ],
        "title": "UserWithStats",
        "description": "User with follow statistics"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}